vorzeitige Anpassung des Partymodus
