+++ src/main.cpp
@@ -1,21 +1,51 @@
-#include "Config.h"
+#include <Arduino.h>
+//#include "avr8-stub.h"
+//#include "app_api.h" //only needed with flash breakpoints
+#include <Adafruit_NeoPixel.h>
+#include <Adafruit_NeoMatrix.h>
+#include <avr/power.h>
+
+#include <Wire.h> //wird gebraucht für I2C-Kommunikation mit dem Gestensensor
+#include "RevEng_PAJ7620.h"
+
+#include "Nextion.h"
+
+// Variablen:
+#define LED_PIN_IndLi    6    // LED Pin für die indirekte Beleuchtung auf der linken Seite an Pin 6
+#define LED_COUNT_IndLi 29    //Anzahl einzelner Neopixel (RGB-LEDs) des LED-Streifens indirekte Beleuchtung auf der linken Seite
+#define LED_PIN_IndRe    5    // LED Pin für die indirekte Beleuchtung auf der linken Seite an Pin 6
+#define LED_COUNT_IndRe 14    //Anzahl einzelner Neopixel (RGB-LEDs) des LED-Streifens indirekte Beleuchtung auf der linken Seite
+#define main_light_width 5
+#define main_light_high 5
+#define main_light_pin 3
+
+#define PIN_LED 13
+#define PIN_schalter 7
 
 int helligkeit; //Wird benötigt für die LDR-Messung
 bool flanke_Licht_ein = false;
 int modus=0;  //1 Lernen, 2 Relax, 3 Mix, 4 Party, 6 Auto
+int modNeu=0;
 int leuchtstaerke=0;
 bool hauptleuchte_an=false;
 bool indirektebeleuchtung_an=false;
 int durchlaufzaehler_party_farbwechsel=0;
 volatile int flankenzaehler_ein_aus=0;
 int activeLamp=0; //0 beide aus; 1 Haupt; 2 Neben; 3 beide
+int parRed[12]={128, 192, 239, 256, 239, 192, 128, 64, 17, 0, 17, 64};
+int parBlue[13]={213, 159, 97, 43, 8, 1, 23, 69, 128, 187, 233, 255, 248};
+int parGreen[14]={2, 5, 32, 78, 134, 189, 232, 254, 251, 224, 178, 122, 67, 24};
+int parBright[11]={128, 169, 196, 202, 185, 149, 107, 71, 54, 60, 87};
+int count=0;
 
-struct lichter[
+/*
+struct lichter{
   int red,green,blue,bright;
-];
+};//0 Aktive Farbe, 1 Lernen, 2 Relax, 3 Mix, 4 Party, 6 Auto
 
-struct lichter Modi[6];
-//int red,green,blue,bright;
+struct lichter Modi[7];
+*/
+int red,green,blue,bright;
 uint32_t memory;
 
 
@@ -26,6 +56,8 @@ int Gestensensor(); //Gestensensor
 int LDR_Messung(); //LDR Messung zwischen 0 und 1023
 void Serielle_Textausgabe(String, String); //Textausgabe zum HMI
 void HMI_Input_loeschen(char*);
+void setModusActive();
+void partymodus();
 
 
 // Objekte:
@@ -43,6 +75,162 @@ Adafruit_NeoMatrix main_light(main_light_width, main_light_high,main_light_pin, 
 
 //Gestensensorobjekt
 RevEng_PAJ7620 sensor = RevEng_PAJ7620();
+
+// TODO: Alle Displayelemente und die Funktionen extrahieren
+
+//Displayelemente
+NexSlider h_red_ls = NexSlider(7, 1, "h_red_ls"); //Slider initialisieren rot; Touch-Release Event muss noch konfiguriert werden
+NexSlider h_green_ls = NexSlider(7, 4, "h_green_ls"); //Slider gruen
+NexSlider h_blue_ls = NexSlider(7, 6, "h_blue_ls"); //Slider blau
+NexSlider h_hell_ls = NexSlider(7, 8, "h3");  //Slider helligkeit
+NexDSButton bt_save_ls = NexDSButton(7, 2, "bt_save_ls"); //Button Licht an/aus
+NexButton b_switch_ls = NexButton(7,15,"b_switch_ls");  //Button um die Lampen umzuschalten
+NexDSButton bt_lernen_lk = NexDSButton(1,13,"bt_lernen_lk");
+NexDSButton bt_relax_lk = NexDSButton(1,15,"bt_relax_lk");
+NexDSButton bt_auto_lk = NexDSButton(1,14,"bt_auto_lk");
+NexDSButton bt_party_lk = NexDSButton(1,16,"bt_party_lk");
+NexDSButton bt_mix_lk = NexDSButton(1,3,"bt_mix_lk");
+NexRadio r_hauptle_lk = NexRadio(1,10,"r_hauptle_lk");
+NexRadio r_indirektb_lk = NexRadio(1,11,"r_indirektb_lk");
+
+NexTouch *nex_listen_list[]={
+  &h_red_ls,
+  &h_green_ls,
+  &h_blue_ls,
+  &h_hell_ls,
+  &bt_save_ls,
+  &b_switch_ls,
+  &bt_lernen_lk,
+  &bt_relax_lk,
+  &bt_auto_lk,
+  &bt_party_lk,
+  &bt_mix_lk,
+  &r_hauptle_lk,
+  &r_indirektb_lk
+};
+
+#pragma region DisplayFunctions
+bool sendCmdToDisplay(String command){
+  Serial.print(command);
+  Serial.write(0xFF);
+  Serial.write(0xFF);
+  Serial.write(0xFF);
+}
+
+void setColorsActive(){
+  if (activeLamp==2)
+  {
+    main_light.fill(main_light.Color(red, green, blue));
+    main_light.setBrightness(bright);
+    main_light.show();
+  }else /*if (activeLamp==3)*/{
+    strip_IndLi.fill(strip_IndLi.Color(red, green, blue));
+    strip_IndLi.setBrightness(bright);
+    strip_IndLi.show();
+    strip_IndRe.fill(strip_IndRe.Color(red, green, blue));
+    strip_IndRe.setBrightness(bright);
+    strip_IndRe.show();
+  }
+}
+
+/*
+void h0PushCallback(void *ptr)      fuer Touch Press Event*/
+
+void h_red_lsPopCallback(void *ptr){
+  h_red_ls.getValue(&memory);
+  red=memory;
+  setColorsActive();
+}
+
+void h_green_lsPopCallback(void *ptr){
+  h_green_ls.getValue(&memory);
+  green=memory;
+  setColorsActive();
+}
+
+void h_blue_lsPopCallback(void *ptr){
+  h_blue_ls.getValue(&memory);
+  blue=memory;
+  setColorsActive();
+}
+
+void h_hell_lsPopCallback(void *ptr){
+  h_hell_ls.getValue(&memory);
+  bright=memory;
+  setColorsActive();
+}
+
+// TODO - Die Farbwerte für den aktuellen Modus speichern
+void bt_save_lsPopCallback(void *ptr){
+  uint32_t state=0;
+  bt_save_ls.getValue(&state);
+
+  if(state == 1){
+    strip_IndLi.fill(strip_IndLi.Color(red, green, blue));
+    strip_IndLi.setBrightness(bright);
+    strip_IndLi.show();
+  }else{
+    strip_IndLi.setBrightness(0);
+    strip_IndLi.show();
+  }
+
+}
+
+void b_switch_lsPopCallback(void *ptr){
+  String text;
+  switch(activeLamp){
+    case 0: activeLamp=1;
+       text="Haupt-\rleuchte";
+       break;
+    case 1: activeLamp=2;
+       text="Neben-\rleuchte";
+       break;
+    case 2: activeLamp=3;
+       text="Beide\rLeuchten";
+       break;
+    default: activeLamp=0;
+       text="Beide\raus";
+       break;
+  }
+  Serielle_Textausgabe("b_switch_ls.txt=",text);
+  setColorsActive();
+}
+
+void bt_lernen_lkPopCallback(void *ptr){
+  modus=1; //Arbeitslicht (Lernen)
+  Serielle_Textausgabe("t_actmod_lk.txt=","Lernen"); //Ausgabetext in Textbox 1 auf der Seite Licht konfiguration (Seite2)
+}
+
+void bt_relax_lkPopCallback(void *ptr){
+  modus=2; //Entspannungslicht (Relax)
+  Serielle_Textausgabe("t_actmod_lk.txt=","Relax"); //Ausgabetext in Textbox 1 auf der Seite Licht konfiguration (Seite2)
+}
+
+void bt_party_lkPopCallback(void *ptr){
+  modus=4; //Party
+  Serielle_Textausgabe("t_actmod_lk.txt=","Party"); //Ausgabetext in Textbox 1 auf der Seite Licht konfiguration (Seite2)
+}
+
+void bt_auto_lkPopCallback(void *ptr){
+  modus=6; //Lichtabhängige Lichtansteuerung (Automatik)
+  Serielle_Textausgabe("t_actmod_lk.txt=","Automatik"); //Ausgabetext in Textbox 1 auf der Seite Licht konfiguration (Seite2)
+}
+
+void bt_mix_lkPopCallback(void *ptr){
+  modus=3; //Farben mix
+  Serielle_Textausgabe("t_actmod_lk.txt=","Farbenmix"); //Ausgabetext in Textbox 1 auf der Seite Licht konfiguration (Seite2)
+}
+
+void r_hauptle_lkPopCallback(void *ptr){
+  r_hauptle_lk.getValue(&memory);
+  hauptleuchte_an=memory?true:false;
+}
+
+void r_indirektb_lkPopCallback(void *ptr){
+  r_indirektb_lk.getValue(&memory);
+  indirektebeleuchtung_an=memory?true:false;
+}
+#pragma endregion
 
 
 // Setupmethode, diese Methode beeinhaltet alle Grundeinstellungen z.B. ob ein Kanal ein Eingang oder Ausgang ist. 
@@ -76,7 +264,7 @@ void setup()
   Serial.println("Please input your gestures:"); */
 
   //Modi initialisieren
-  Modi[0]
+  //1 Lernen, 2 Relax, 3 Mix, 4 Party, 6 Auto
 
 
   //Display Events zu Artefakten hinzufuegen
@@ -111,7 +299,13 @@ void setup()
 void loop() 
 {
   nexLoop(nex_listen_list);
+  if(modNeu!=modus){
+    setModusActive();
+  }
 
+  if(modus==4){
+    partymodus();
+  }
   // TODO: Je nach aktivem Modus die Lampen ansteuern?
   //RGB_Licht_Funktion(0, 0, 0, 0, 255, 4);
   //RGB_Licht_Funktion(0, 0, 0, 0, 255, 6);
@@ -134,21 +328,22 @@ void loop()
 }
 
 void setModusActive(){
-  switch (modus){
+  switch (modNeu){
     case 1: //Lernen
       red=255;
       green=255;
       blue=255;
-      leuchtstaerke=255;
+      bright=255;
       break;
     case 2: //Entspannungslicht (Relax)
       red=241;
       green=142;
       blue=28;
-      leuchtstaerke=255;
+      bright=255;
       break;
     case 4: //Party
-      leuchtstaerke=255;
+        //wird in anderen Funktion gehändelt
+        //Task aktivieren?
       break;
     case 3: //Farben mix
       break;
@@ -157,128 +352,27 @@ void setModusActive(){
     default:
       break;
   }
+  //alten Modus abschalten
+  switch(modus){
+    
+  }
 }
 
+void partymodus(){
+  red=parRed[count%12];
+  blue=parBlue[count%13];
+  green=parGreen[count%14];
+  bright=parBright[count%11];
+  count++;
+  delay(50);
+}
 
 int LDR_Messung()
 {
   helligkeit = analogRead(0);
   Serial.println(helligkeit);
   delay(500);
-  
-
-/*
-   // das Signal wird nur aller 5 Sekunden abgefragt
-  delay(950);
-
-
-  if(hmi_input[1]==1 && hmi_input[2]==0x0F && indirektebeleuchtung_an==false) //Auswertung welche Leuchten angesteuert werden sollen
-  {
-    indirektebeleuchtung_an=true;
-    HMI_Input_loeschen(hmi_input);
-  }
-  else if(hmi_input[1]==1 && hmi_input[2]==0x0F && indirektebeleuchtung_an==true)
-  {
-    indirektebeleuchtung_an=false;
-    HMI_Input_loeschen(hmi_input);
-  }
-  if(hmi_input[1]==1 && hmi_input[2]==0x0E && hauptleuchte_an==false)
-  {
-    hauptleuchte_an=true;
-    HMI_Input_loeschen(hmi_input);
-  }
-  else if(hmi_input[1]==1 && hmi_input[2]==0x0E && hauptleuchte_an==true)
-  {
-    hauptleuchte_an=false;
-    HMI_Input_loeschen(hmi_input);
-  }
-
-  if(hmi_input[1]==0 && hmi_input[2]==5 && modus!=0)  //Ein/Aus der ausgewählten Leuchten
-  {
-    flankenzaehler_ein_aus++;
-    if(flankenzaehler_ein_aus==1)
-    {
-      flanke_Licht_ein=true;
-      Serielle_Textausgabe("texbox_1.txt=","Licht Ein"); //Ausgabetext in Textbox 1 auf Seite 1
-      if(modus!=4 || modus !=6){
-        RGB_Licht_Funktion(0, rot, gruen, blau, leuchtstaerke, modus, hauptleuchte_an, indirektebeleuchtung_an);
-      }  
-    } 
-    else if(flankenzaehler_ein_aus==2)
-    {
-      flanke_Licht_ein=false;
-      flankenzaehler_ein_aus=0;
-      Serielle_Textausgabe("texbox_1.txt=","Licht Aus"); //Ausgabetext in Textbox 1 auf Seite 1 
-      RGB_Licht_Funktion(0, 0, 0, 0, 0, modus, hauptleuchte_an, indirektebeleuchtung_an);
-    }
-     HMI_Input_loeschen(hmi_input);
-  }
-   
-
-  if(modus==4 && flanke_Licht_ein==true) //Farbwechsel für Partymodus
-  {
-    if(durchlaufzaehler_party_farbwechsel<=100)
-    {
-      rot=255;
-      gruen=0;
-      blau=0;
-    }
-    else if(durchlaufzaehler_party_farbwechsel<=200)
-    {
-      rot=0;
-      gruen=255;
-      blau=0;
-    }
-    else if(durchlaufzaehler_party_farbwechsel<=300)
-    {
-      rot=0;
-      gruen=0;
-      blau=255;
-    }
-    else if(durchlaufzaehler_party_farbwechsel==301)
-    {
-      durchlaufzaehler_party_farbwechsel=0;
-    }
-    durchlaufzaehler_party_farbwechsel++;
-    RGB_Licht_Funktion(0, rot, gruen, blau, 255, modus, hauptleuchte_an, indirektebeleuchtung_an);
-  }
-
-
-  if(modus==6 && flanke_Licht_ein==true) //Automatikmodus
-  {
-    int aktueller_LDR_wert=LDR_Messung();
-    int setze_helligkeit;
-
-      if(aktueller_LDR_wert<=300)
-      {
-        setze_helligkeit=255;
-      }
-      else if (aktueller_LDR_wert<=500)
-      {
-        setze_helligkeit=200;
-      }
-      else if (aktueller_LDR_wert<=700)
-      {
-        setze_helligkeit=150;
-      }
-      else if (aktueller_LDR_wert<=850)
-      {
-        setze_helligkeit=50;
-      }
-      else if (aktueller_LDR_wert<=1000)
-      {
-        setze_helligkeit=0;
-      }
-    rot=241;
-    gruen=142;
-    blau=28;
-    RGB_Licht_Funktion(0, rot, gruen, blau, setze_helligkeit, modus, hauptleuchte_an, indirektebeleuchtung_an);
-  }
- 
-  
-*/
 }
-
 
 void Serielle_Textausgabe(String textbox, String text)
 {
@@ -291,14 +385,6 @@ void Serielle_Textausgabe(String textbox, String text)
       Serial.write(0xFF);
   }
 }
-/*
-int LDR_Messung()
-{
-  helligkeit = analogRead(0);  
-  return helligkeit;
-}
-*/
-
 
 int Gestensensor()
 {
